backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pyyaml

# Testing
pytest
pytest-asyncio
httpx
